dnsdist: Split the Lua files, remove deprecated functions
