+++ pdns/dnsdist-lua.cc
@@ -19,19 +19,25 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
+
+#include <dirent.h>
+#include <fstream>
+#include <net/if.h>
 #include <sys/types.h>
 #include <sys/socket.h>
-#include <net/if.h>
-#include "dnsdist.hh"
-#include "dnsrulactions.hh"
+#include <sys/stat.h>
 #include <thread>
-#include "dolog.hh"
-#include "sodcrypto.hh"
-#include "base64.hh"
-#include <fstream>
-#include "dnswriter.hh"
-#include "lock.hh"
+
+#include "dnsdist.hh"
 #include "dnsdist-lua.hh"
+
+#include "base64.hh"
+#include "dnswriter.hh"
+#include "dolog.hh"
+#include "lock.hh"
+#include "sodcrypto.hh"
+
+#include <boost/logic/tribool.hpp>
 
 #ifdef HAVE_SYSTEMD
 #include <systemd/sd-daemon.h>
@@ -39,145 +45,39 @@
 
 using std::thread;
 
-static vector<std::function<void(void)>>* g_launchWork;
+static vector<std::function<void(void)>>* g_launchWork = nullptr;
 
-class LuaAction : public DNSAction
+boost::tribool g_noLuaSideEffect;
+static bool g_included{false};
+
+/* this is a best effort way to prevent logging calls with no side-effects in the output of delta()
+   Functions can declare setLuaNoSideEffect() and if nothing else does declare a side effect, or nothing
+   has done so before on this invocation, this call won't be part of delta() output */
+void setLuaNoSideEffect()
 {
-public:
-  typedef std::function<std::tuple<int, string>(DNSQuestion* dq)> func_t;
-  LuaAction(LuaAction::func_t func) : d_func(func)
-  {}
-
-  Action operator()(DNSQuestion* dq, string* ruleresult) const override
-  {
-    std::lock_guard<std::mutex> lock(g_luamutex);
-    auto ret = d_func(dq);
-    if(ruleresult)
-      *ruleresult=std::get<1>(ret);
-    return (Action)std::get<0>(ret);
-  }
-
-  string toString() const override
-  {
-    return "Lua script";
-  }
-
-private:
-  func_t d_func;
-};
-
-class LuaResponseAction : public DNSResponseAction
-{
-public:
-  typedef std::function<std::tuple<int, string>(DNSResponse* dr)> func_t;
-  LuaResponseAction(LuaResponseAction::func_t func) : d_func(func)
-  {}
-
-  Action operator()(DNSResponse* dr, string* ruleresult) const override
-  {
-    std::lock_guard<std::mutex> lock(g_luamutex);
-    auto ret = d_func(dr);
-    if(ruleresult)
-      *ruleresult=std::get<1>(ret);
-    return (Action)std::get<0>(ret);
-  }
-
-  string toString() const override
-  {
-    return "Lua response script";
-  }
-
-private:
-  func_t d_func;
-};
-
-std::shared_ptr<DNSRule> makeRule(const luadnsrule_t& var)
-{
-  if (var.type() == typeid(std::shared_ptr<DNSRule>))
-    return *boost::get<std::shared_ptr<DNSRule>>(&var);
-
-  SuffixMatchNode smn;
-  NetmaskGroup nmg;
-  auto add=[&](string src) {
-    try {
-      nmg.addMask(src); // need to try mask first, all masks are domain names!
-    } catch(...) {
-      smn.add(DNSName(src));
-    }
-  };
-
-  if (var.type() == typeid(string))
-    add(*boost::get<string>(&var));
-
-  else if (var.type() == typeid(vector<pair<int, string>>))
-    for(const auto& a : *boost::get<vector<pair<int, string>>>(&var))
-      add(a.second);
-
-  else if (var.type() == typeid(DNSName))
-    smn.add(*boost::get<DNSName>(&var));
-
-  else if (var.type() == typeid(vector<pair<int, DNSName>>))
-    for(const auto& a : *boost::get<vector<pair<int, DNSName>>>(&var))
-      smn.add(a.second);
-
-  if(nmg.empty())
-    return std::make_shared<SuffixMatchNodeRule>(smn);
-  else
-    return std::make_shared<NetmaskGroupRule>(nmg, true);
+  if(g_noLuaSideEffect==false) // there has been a side effect already
+    return;
+  g_noLuaSideEffect=true;
 }
 
-std::unordered_map<int, vector<boost::variant<string,double>>> getGenResponses(unsigned int top, boost::optional<int> labels, std::function<bool(const Rings::Response&)> pred) 
+void setLuaSideEffect()
 {
-  setLuaNoSideEffect();
-  map<DNSName, int> counts;
-  unsigned int total=0;
-  {
-    std::lock_guard<std::mutex> lock(g_rings.respMutex);
-    if(!labels) {
-      for(const auto& a : g_rings.respRing) {
-        if(!pred(a))
-          continue;
-        counts[a.name]++;
-        total++;
-      }
-    }
-    else {
-      unsigned int lab = *labels;
-      for(auto a : g_rings.respRing) {
-        if(!pred(a))
-          continue;
-        
-        a.name.trimToLabels(lab);
-        counts[a.name]++;
-        total++;
-      }
-      
-    }
-  }
-  //      cout<<"Looked at "<<total<<" responses, "<<counts.size()<<" different ones"<<endl;
-  vector<pair<int, DNSName>> rcounts;
-  rcounts.reserve(counts.size());
-  for(const auto& c : counts) 
-    rcounts.push_back(make_pair(c.second, c.first.makeLowerCase()));
-  
-  sort(rcounts.begin(), rcounts.end(), [](const decltype(rcounts)::value_type& a, 
-                                          const decltype(rcounts)::value_type& b) {
-         return b.first < a.first;
-       });
-  
-  std::unordered_map<int, vector<boost::variant<string,double>>> ret;
-  unsigned int count=1, rest=0;
-  for(const auto& rc : rcounts) {
-    if(count==top+1)
-      rest+=rc.first;
-    else
-      ret.insert({count++, {rc.second.toString(), rc.first, 100.0*rc.first/total}});
-  }
-  ret.insert({count, {"Rest", rest, total > 0 ? 100.0*rest/total : 100.0}});
-  return ret;
+  g_noLuaSideEffect=false;
 }
 
-void parseLocalBindVars(boost::optional<localbind_t> vars, bool& doTCP, bool& reusePort, int& tcpFastOpenQueueSize, std::string& interface, std::set<int>& cpus)
+bool getLuaNoSideEffect()
+{
+  return g_noLuaSideEffect==true;
+}
+
+void resetLuaSideEffect()
+{
+  g_noLuaSideEffect = boost::logic::indeterminate;
+}
+
+typedef std::unordered_map<std::string, boost::variant<bool, int, std::string, std::vector<std::pair<int,int> > > > localbind_t;
+
+static void parseLocalBindVars(boost::optional<localbind_t> vars, bool& doTCP, bool& reusePort, int& tcpFastOpenQueueSize, std::string& interface, std::set<int>& cpus)
 {
   if (vars) {
     if (vars->count("doTCP")) {
@@ -200,80 +100,16 @@ void parseLocalBindVars(boost::optional<localbind_t> vars, bool& doTCP, bool& re
   }
 }
 
-vector<std::function<void(void)>> setupLua(bool client, const std::string& config)
+void setupLuaConfig(bool client)
 {
-  g_launchWork= new vector<std::function<void(void)>>();
   typedef std::unordered_map<std::string, boost::variant<bool, std::string, vector<pair<int, std::string> > > > newserver_t;
 
-  g_lua.writeVariable("DNSAction", std::unordered_map<string,int>{
-      {"Drop", (int)DNSAction::Action::Drop},
-      {"Nxdomain", (int)DNSAction::Action::Nxdomain},
-      {"Refused", (int)DNSAction::Action::Refused},
-      {"Spoof", (int)DNSAction::Action::Spoof},
-      {"Allow", (int)DNSAction::Action::Allow},
-      {"HeaderModify", (int)DNSAction::Action::HeaderModify},
-      {"Pool", (int)DNSAction::Action::Pool},
-      {"None",(int)DNSAction::Action::None},
-      {"Delay", (int)DNSAction::Action::Delay},
-      {"Truncate", (int)DNSAction::Action::Truncate}
-    });
-
-  g_lua.writeVariable("DNSResponseAction", std::unordered_map<string,int>{
-      {"Allow",        (int)DNSResponseAction::Action::Allow        },
-      {"Delay",        (int)DNSResponseAction::Action::Delay        },
-      {"HeaderModify", (int)DNSResponseAction::Action::HeaderModify },
-      {"None",         (int)DNSResponseAction::Action::None         }
-    });
-
-  g_lua.writeVariable("DNSClass", std::unordered_map<string,int>{
-      {"IN",    QClass::IN    },
-      {"CHAOS", QClass::CHAOS },
-      {"NONE",  QClass::NONE  },
-      {"ANY",   QClass::ANY   }
-    });
-
-  g_lua.writeVariable("DNSOpcode", std::unordered_map<string,int>{
-      {"Query",  Opcode::Query  },
-      {"IQuery", Opcode::IQuery },
-      {"Status", Opcode::Status },
-      {"Notify", Opcode::Notify },
-      {"Update", Opcode::Update }
-    });
-
-  g_lua.writeVariable("DNSSection", std::unordered_map<string,int>{
-      {"Question",  0 },
-      {"Answer",    1 },
-      {"Authority", 2 },
-      {"Additional",3 }
-    });
-
-  vector<pair<string, int> > rcodes = {{"NOERROR",  RCode::NoError  },
-                                       {"FORMERR",  RCode::FormErr  },
-                                       {"SERVFAIL", RCode::ServFail },
-                                       {"NXDOMAIN", RCode::NXDomain },
-                                       {"NOTIMP",   RCode::NotImp   },
-                                       {"REFUSED",  RCode::Refused  },
-                                       {"YXDOMAIN", RCode::YXDomain },
-                                       {"YXRRSET",  RCode::YXRRSet  },
-                                       {"NXRRSET",  RCode::NXRRSet  },
-                                       {"NOTAUTH",  RCode::NotAuth  },
-                                       {"NOTZONE",  RCode::NotZone  }
-  };
-  vector<pair<string, int> > dd;
-  for(const auto& n : QType::names)
-    dd.push_back({n.first, n.second});
-  for(const auto& n : rcodes)
-    dd.push_back({n.first, n.second});
-  g_lua.writeVariable("dnsdist", dd);
-  
   g_lua.writeFunction("inClientStartup", [client]() {
         return client && !g_configurationDone;
   });
 
-
-  g_lua.writeFunction("newServer", 
-		      [client](boost::variant<string,newserver_t> pvars, boost::optional<int> qps)
-		      { 
+  g_lua.writeFunction("newServer",
+		      [client](boost::variant<string,newserver_t> pvars, boost::optional<int> qps) {
                         setLuaSideEffect();
 			if(client) {
 			  return std::make_shared<DownstreamState>(ComboAddress());
@@ -306,8 +142,8 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
 			  if(qps) {
 			    ret->qps=QPSLimiter(*qps, *qps);
 			  }
-			  g_dstates.modify([ret](servers_t& servers) { 
-			      servers.push_back(ret); 
+			  g_dstates.modify([ret](servers_t& servers) {
+			      servers.push_back(ret);
 			      std::stable_sort(servers.begin(), servers.end(), [](const decltype(ret)& a, const decltype(ret)& b) {
 				  return a->order < b->order;
 				});
@@ -530,88 +366,9 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
 			return ret;
 		      } );
 
-  g_lua.writeFunction("makeRule", makeRule);
-
-  g_lua.writeFunction("addAnyTCRule", []() {
-      setLuaSideEffect();
-      warnlog("addAnyTCRule() is deprecated and will be removed in 1.3.0, please use addAction(AndRule{QTypeRule(dnsdist.ANY), TCPRule(false)}, TCAction()) instead");
-
-      auto rules=g_rulactions.getCopy();
-      std::vector<pair<int, shared_ptr<DNSRule> >> v;
-      v.push_back({1, std::make_shared<QTypeRule>(0xff)});
-      v.push_back({2, std::make_shared<TCPRule>(false)});
-      rules.push_back({ std::shared_ptr<DNSRule>(new AndRule(v)), std::make_shared<TCAction>()});
-      g_rulactions.setState(rules);
-    });
-
-  g_lua.writeFunction("rmRule", [](unsigned int num) {
-      setLuaSideEffect();
-      auto rules = g_rulactions.getCopy();
-      if(num >= rules.size()) {
-	g_outputBuffer = "Error: attempt to delete non-existing rule\n";
-	return;
-      }
-      rules.erase(rules.begin()+num);
-      g_rulactions.setState(rules);
-    });
-
-  g_lua.writeFunction("topRule", []() {
-      setLuaSideEffect();
-      auto rules = g_rulactions.getCopy();
-      if(rules.empty())
-	return;
-      auto subject = *rules.rbegin();
-      rules.erase(std::prev(rules.end()));
-      rules.insert(rules.begin(), subject);
-      g_rulactions.setState(rules);
-    });
-  g_lua.writeFunction("mvRule", [](unsigned int from, unsigned int to) {
-      setLuaSideEffect();
-      auto rules = g_rulactions.getCopy();
-      if(from >= rules.size() || to > rules.size()) {
-	g_outputBuffer = "Error: attempt to move rules from/to invalid index\n";
-	return;
-      }
-
-      auto subject = rules[from];
-      rules.erase(rules.begin()+from);
-      if(to == rules.size())
-	rules.push_back(subject);
-      else {
-	if(from < to)
-	  --to;
-	rules.insert(rules.begin()+to, subject);
-      }
-      g_rulactions.setState(rules);
-    });
-  g_lua.writeFunction("clearRules", []() {
-      setLuaSideEffect();
-      g_rulactions.modify([](decltype(g_rulactions)::value_type& rulactions) {
-          rulactions.clear();
-        });
-    });
-
-  g_lua.writeFunction("newRuleAction", [](luadnsrule_t dnsrule, std::shared_ptr<DNSAction> action) {
-      auto rule=makeRule(dnsrule);
-      return std::make_shared<std::pair< luadnsrule_t, std::shared_ptr<DNSAction> > >(rule, action);
-    });
-
-  g_lua.writeFunction("setRules", [](std::vector< std::pair<int, std::shared_ptr<std::pair<luadnsrule_t, std::shared_ptr<DNSAction> > > > > newruleactions) {
-      setLuaSideEffect();
-      g_rulactions.modify([newruleactions](decltype(g_rulactions)::value_type& gruleactions) {
-          gruleactions.clear();
-          for (const auto& newruleaction : newruleactions) {
-            if (newruleaction.second) {
-              auto rule=makeRule(newruleaction.second->first);
-              gruleactions.push_back({rule, newruleaction.second->second});
-            }
-          }
-        });
-    });
-
-  g_lua.writeFunction("rmServer", 
+  g_lua.writeFunction("rmServer",
 		      [](boost::variant<std::shared_ptr<DownstreamState>, int> var)
-		      { 
+		      {
                         setLuaSideEffect();
                         shared_ptr<DownstreamState> server;
                         auto* rem = boost::get<shared_ptr<DownstreamState>>(&var);
@@ -634,7 +391,6 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
                         g_dstates.setState(states);
 		      } );
 
-
   g_lua.writeFunction("setServerPolicy", [](ServerPolicy policy)  {
       setLuaSideEffect();
       g_policy.setState(policy);
@@ -652,14 +408,6 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
   g_lua.writeFunction("truncateTC", [](bool tc) { setLuaSideEffect(); g_truncateTC=tc; });
   g_lua.writeFunction("fixupCase", [](bool fu) { setLuaSideEffect(); g_fixupCase=fu; });
 
-  g_lua.registerMember("name", &ServerPolicy::name);
-  g_lua.registerMember("policy", &ServerPolicy::policy);
-  g_lua.writeFunction("newServerPolicy", [](string name, policyfunc_t policy) { return ServerPolicy{name, policy};});
-  g_lua.writeVariable("firstAvailable", ServerPolicy{"firstAvailable", firstAvailable});
-  g_lua.writeVariable("roundrobin", ServerPolicy{"roundrobin", roundrobin});
-  g_lua.writeVariable("wrandom", ServerPolicy{"wrandom", wrandom});
-  g_lua.writeVariable("whashed", ServerPolicy{"whashed", whashed});
-  g_lua.writeVariable("leastOutstanding", ServerPolicy{"leastOutstanding", leastOutstanding});
   g_lua.writeFunction("addACL", [](const std::string& domain) {
       setLuaSideEffect();
       g_ACL.modify([domain](NetmaskGroup& nmg) { nmg.addMask(domain); });
@@ -715,6 +463,7 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
 	g_outputBuffer="Error: "+string(e.what())+"\n";
       }
     });
+
   g_lua.writeFunction("setACL", [](boost::variant<string,vector<pair<int, string>>> inp) {
       setLuaSideEffect();
       NetmaskGroup nmg;
@@ -726,6 +475,7 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
       }
       g_ACL.setState(nmg);
   });
+
   g_lua.writeFunction("showACL", []() {
       setLuaNoSideEffect();
       vector<string> vec;
@@ -736,6 +486,7 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
         g_outputBuffer+=s+"\n";
 
     });
+
   g_lua.writeFunction("shutdown", []() {
 #ifdef HAVE_SYSTEMD
       sd_notify(0, "STOPPING=1");
@@ -743,20 +494,7 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
       _exit(0);
   } );
 
-
-  g_lua.writeFunction("addDomainBlock", [](const std::string& domain) { 
-      setLuaSideEffect();
-      warnlog("addDomainBlock() is deprecated and will be removed in 1.3.0, please use addAction(\"%s\", DropAction()) instead", domain);
-      SuffixMatchNode smn;
-      smn.add(DNSName(domain));
-	g_rulactions.modify([smn](decltype(g_rulactions)::value_type& rulactions) {
-	    rulactions.push_back({
-				   std::make_shared<SuffixMatchNodeRule>(smn), 
-				   std::make_shared<DropAction>()  });
-	  });
-
-    });
-  g_lua.writeFunction("showServers", []() {  
+  g_lua.writeFunction("showServers", []() {
       setLuaNoSideEffect();
       try {
       ostringstream ret;
@@ -777,7 +515,7 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
 	}
 
 	ret << (fmt % counter % s->name % s->remote.toStringWithPort() %
-		status % 
+		status %
 		s->queryLoad % s->qps.getRate() % s->order % s->weight % s->queries.load() % s->reuseds.load() % (s->dropRate) % (s->latencyUsec/1000.0) % s->outstanding.load() % pools) << endl;
 
 	totQPS += s->queryLoad;
@@ -786,421 +524,11 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
 	++counter;
       }
       ret<< (fmt % "All" % "" % "" % ""
-		% 
+		%
 	     (double)totQPS % "" % "" % "" % totQueries % totDrops % "" % "" % "" % "" ) << endl;
 
       g_outputBuffer=ret.str();
       }catch(std::exception& e) { g_outputBuffer=e.what(); throw; }
-    });
-
-  g_lua.writeFunction("addLuaAction", [](luadnsrule_t var, LuaAction::func_t func) 
-		      {
-                        setLuaSideEffect();
-			auto rule=makeRule(var);
-			g_rulactions.modify([rule,func](decltype(g_rulactions)::value_type& rulactions){
-			    rulactions.push_back({rule,
-				  std::make_shared<LuaAction>(func)});
-			  });
-		      });
-
-  g_lua.writeFunction("LuaAction", [](LuaAction::func_t func) {
-      setLuaSideEffect();
-      return std::shared_ptr<DNSAction>(new LuaAction(func));
-    });
-
-  g_lua.writeFunction("addLuaResponseAction", [](luadnsrule_t var, LuaResponseAction::func_t func) {
-      setLuaSideEffect();
-      auto rule=makeRule(var);
-      g_resprulactions.modify([rule,func](decltype(g_resprulactions)::value_type& rulactions){
-          rulactions.push_back({rule,
-                std::make_shared<LuaResponseAction>(func)});
-        });
-    });
-
-  g_lua.writeFunction("LuaResponseAction", [](LuaResponseAction::func_t func) {
-      setLuaSideEffect();
-      return std::shared_ptr<DNSResponseAction>(new LuaResponseAction(func));
-    });
-
-  g_lua.writeFunction("NoRecurseAction", []() {
-      return std::shared_ptr<DNSAction>(new NoRecurseAction);
-    });
-
-  g_lua.writeFunction("MacAddrAction", [](int code) {
-      return std::shared_ptr<DNSAction>(new MacAddrAction(code));
-    });
-
-
-  g_lua.writeFunction("PoolAction", [](const string& a) {
-      return std::shared_ptr<DNSAction>(new PoolAction(a));
-    });
-
-  g_lua.writeFunction("QPSAction", [](int limit) {
-      return std::shared_ptr<DNSAction>(new QPSAction(limit));
-    });
-
-  g_lua.writeFunction("QPSPoolAction", [](int limit, const string& a) {
-      return std::shared_ptr<DNSAction>(new QPSPoolAction(limit, a));
-    });
-
-  g_lua.writeFunction("SpoofAction", [](boost::variant<string,vector<pair<int, string>>> inp, boost::optional<string> b ) {
-      vector<ComboAddress> addrs;
-      if(auto s = boost::get<string>(&inp))
-        addrs.push_back(ComboAddress(*s));
-      else {
-        const auto& v = boost::get<vector<pair<int,string>>>(inp);
-        for(const auto& a: v)
-          addrs.push_back(ComboAddress(a.second));
-      }
-      if(b)
-        addrs.push_back(ComboAddress(*b));
-      return std::shared_ptr<DNSAction>(new SpoofAction(addrs));
-    });
-
-  g_lua.writeFunction("SpoofCNAMEAction", [](const string& a) {
-      return std::shared_ptr<DNSAction>(new SpoofAction(a));
-    });
-
-  g_lua.writeFunction("addDomainSpoof", [](const std::string& domain, boost::variant<string,vector<pair<int, string>>> inp, boost::optional<string> b) { 
-      setLuaSideEffect();
-      warnlog("addDomainSpoof() is deprecated and will be removed in 1.3.0, please use addAction(\"%s\", SpoofAction(...)) instead", domain);
-
-      SuffixMatchNode smn;
-      vector<ComboAddress> outp;
-      try
-      {
-	smn.add(DNSName(domain));
-
-        if(auto s = boost::get<string>(&inp))
-          outp.push_back(ComboAddress(*s));
-        else {
-          const auto& v = boost::get<vector<pair<int,string>>>(inp);
-          for(const auto& a: v)
-            outp.push_back(ComboAddress(a.second));
-        }
-        if(b)
-          outp.push_back(ComboAddress(*b));
-          
-      }
-      catch(std::exception& e) {
-	g_outputBuffer="Error parsing parameters: "+string(e.what());
-	return;
-      }
-      g_rulactions.modify([&smn,&outp](decltype(g_rulactions)::value_type& rulactions) {
-	  rulactions.push_back({
-	      std::make_shared<SuffixMatchNodeRule>(smn), 
-		std::make_shared<SpoofAction>(outp)  });
-	});
-
-    });
-
-  g_lua.writeFunction("addDomainCNAMESpoof", [](const std::string& domain, const std::string& cname) {
-      setLuaSideEffect();
-      warnlog("addDomainCNAMESpoof() is deprecated and will be removed in 1.3.0, please use addAction(\"%s\", SpoofCNAMEAction(\"%s\")) instead", domain, cname);
-
-      SuffixMatchNode smn;
-      try
-      {
-	smn.add(DNSName(domain));
-      }
-      catch(std::exception& e) {
-	g_outputBuffer="Error parsing parameters: "+string(e.what());
-	return;
-      }
-      g_rulactions.modify([&smn,&cname](decltype(g_rulactions)::value_type& rulactions) {
-	  rulactions.push_back({
-	      std::make_shared<SuffixMatchNodeRule>(smn),
-		std::make_shared<SpoofAction>(cname)  });
-	});
-    });
-
-  g_lua.writeFunction("DropAction", []() {
-      return std::shared_ptr<DNSAction>(new DropAction);
-    });
-
-  g_lua.writeFunction("AllowAction", []() {
-      return std::shared_ptr<DNSAction>(new AllowAction);
-    });
-
-  g_lua.writeFunction("DelayAction", [](int msec) {
-      return std::shared_ptr<DNSAction>(new DelayAction(msec));
-    });
-
-  g_lua.writeFunction("TCAction", []() {
-      return std::shared_ptr<DNSAction>(new TCAction);
-    });
-
-  g_lua.writeFunction("DisableValidationAction", []() {
-      return std::shared_ptr<DNSAction>(new DisableValidationAction);
-    });
-
-  g_lua.writeFunction("LogAction", [](const std::string& fname, boost::optional<bool> binary, boost::optional<bool> append, boost::optional<bool> buffered) {
-      return std::shared_ptr<DNSAction>(new LogAction(fname, binary ? *binary : true, append ? *append : false, buffered ? *buffered : false));
-    });
-
-  g_lua.writeFunction("RCodeAction", [](int rcode) {
-      return std::shared_ptr<DNSAction>(new RCodeAction(rcode));
-    });
-
-  g_lua.writeFunction("SkipCacheAction", []() {
-      return std::shared_ptr<DNSAction>(new SkipCacheAction);
-    });
-
-  g_lua.writeFunction("MaxQPSIPRule", [](unsigned int qps, boost::optional<int> ipv4trunc, boost::optional<int> ipv6trunc, boost::optional<int> burst) {
-        return std::shared_ptr<DNSRule>(new MaxQPSIPRule(qps, burst.get_value_or(qps), ipv4trunc.get_value_or(32), ipv6trunc.get_value_or(64)));
-    });
-
-
-  g_lua.writeFunction("MaxQPSRule", [](unsigned int qps, boost::optional<int> burst) {
-      if(!burst)
-        return std::shared_ptr<DNSRule>(new MaxQPSRule(qps));
-      else
-        return std::shared_ptr<DNSRule>(new MaxQPSRule(qps, *burst));      
-    });
-
-
-  g_lua.writeFunction("RegexRule", [](const std::string& str) {
-      return std::shared_ptr<DNSRule>(new RegexRule(str));
-    });
-
-#ifdef HAVE_RE2
-  g_lua.writeFunction("RE2Rule", [](const std::string& str) {
-      return std::shared_ptr<DNSRule>(new RE2Rule(str));
-    });
-#endif
-
-  g_lua.writeFunction("SuffixMatchNodeRule", [](const SuffixMatchNode& smn, boost::optional<bool> quiet) {
-      return std::shared_ptr<DNSRule>(new SuffixMatchNodeRule(smn, quiet ? *quiet : false));
-    });
-
-  g_lua.writeFunction("NetmaskGroupRule", [](const NetmaskGroup& nmg, boost::optional<bool> src) {
-      return std::shared_ptr<DNSRule>(new NetmaskGroupRule(nmg, src ? *src : true));
-    });
-
-  g_lua.writeFunction("benchRule", [](std::shared_ptr<DNSRule> rule, boost::optional<int> times_, boost::optional<string> suffix_)  {
-      setLuaNoSideEffect();
-      int times = times_.get_value_or(100000);
-      DNSName suffix(suffix_.get_value_or("powerdns.com"));
-      struct item {
-        vector<uint8_t> packet;        
-        ComboAddress rem;
-        DNSName qname;
-        uint16_t qtype, qclass;
-      };
-      vector<item> items;
-      items.reserve(1000);
-      for(int n=0; n < 1000; ++n) {
-        struct item i;
-        i.qname=DNSName(std::to_string(random()));
-        i.qname += suffix;
-        i.qtype = random() % 0xff;
-        i.qclass = 1;
-        i.rem=ComboAddress("127.0.0.1");
-        i.rem.sin4.sin_addr.s_addr = random();
-        DNSPacketWriter pw(i.packet, i.qname, i.qtype);
-        items.push_back(i);
-      }
-
-      int matches=0;
-      ComboAddress dummy("127.0.0.1");
-      DTime dt;
-      dt.set();
-      for(int n=0; n < times; ++n) {
-        const item& i = items[n % items.size()];
-        DNSQuestion dq(&i.qname, i.qtype, i.qclass, &i.rem, &i.rem, (struct dnsheader*)&i.packet[0], i.packet.size(), i.packet.size(), false);
-        if(rule->matches(&dq))
-          matches++;
-      }
-      double udiff=dt.udiff();
-      g_outputBuffer=(boost::format("Had %d matches out of %d, %.1f qps, in %.1f usec\n") % matches % times % (1000000*(1.0*times/udiff)) % udiff).str();
-
-    });
-
-  g_lua.writeFunction("AllRule", []() {
-      return std::shared_ptr<DNSRule>(new AllRule());
-    });
-
-  g_lua.writeFunction("ProbaRule", [](double proba) {
-      return std::shared_ptr<DNSRule>(new ProbaRule(proba));
-    });
-
-  
-  g_lua.writeFunction("QNameRule", [](const std::string& qname) {
-      return std::shared_ptr<DNSRule>(new QNameRule(DNSName(qname)));
-    });
-  
-  g_lua.writeFunction("QTypeRule", [](boost::variant<int, std::string> str) {
-      uint16_t qtype;
-      if(auto dir = boost::get<int>(&str)) {
-        qtype = *dir;
-      }
-      else {
-        string val=boost::get<string>(str);
-        qtype = QType::chartocode(val.c_str());
-        if(!qtype)
-          throw std::runtime_error("Unable to convert '"+val+"' to a DNS type");
-      }
-      return std::shared_ptr<DNSRule>(new QTypeRule(qtype));
-    });
-  g_lua.writeFunction("QClassRule", [](int c) {
-      return std::shared_ptr<DNSRule>(new QClassRule(c));
-    });
-
-  g_lua.writeFunction("OpcodeRule", [](uint8_t code) {
-      return std::shared_ptr<DNSRule>(new OpcodeRule(code));
-    });
-
-  g_lua.writeFunction("AndRule", [](vector<pair<int, std::shared_ptr<DNSRule> > >a) {
-      return std::shared_ptr<DNSRule>(new AndRule(a));
-    });
-
-  g_lua.writeFunction("OrRule", [](vector<pair<int, std::shared_ptr<DNSRule> > >a) {
-      return std::shared_ptr<DNSRule>(new OrRule(a));
-    });
-
-  g_lua.writeFunction("TCPRule", [](bool tcp) {
-      return std::shared_ptr<DNSRule>(new TCPRule(tcp));
-    });
-
-  g_lua.writeFunction("DNSSECRule", []() {
-      return std::shared_ptr<DNSRule>(new DNSSECRule());
-    });
-
-  g_lua.writeFunction("NotRule", [](std::shared_ptr<DNSRule>rule) {
-      return std::shared_ptr<DNSRule>(new NotRule(rule));
-    });
-
-  g_lua.writeFunction("RecordsCountRule", [](uint8_t section, uint16_t minCount, uint16_t maxCount) {
-      return std::shared_ptr<DNSRule>(new RecordsCountRule(section, minCount, maxCount));
-    });
-
-  g_lua.writeFunction("RecordsTypeCountRule", [](uint8_t section, uint16_t type, uint16_t minCount, uint16_t maxCount) {
-      return std::shared_ptr<DNSRule>(new RecordsTypeCountRule(section, type, minCount, maxCount));
-    });
-
-  g_lua.writeFunction("TrailingDataRule", []() {
-      return std::shared_ptr<DNSRule>(new TrailingDataRule());
-    });
-
-  g_lua.writeFunction("QNameLabelsCountRule", [](unsigned int minLabelsCount, unsigned int maxLabelsCount) {
-      return std::shared_ptr<DNSRule>(new QNameLabelsCountRule(minLabelsCount, maxLabelsCount));
-    });
-
-  g_lua.writeFunction("QNameWireLengthRule", [](size_t min, size_t max) {
-      return std::shared_ptr<DNSRule>(new QNameWireLengthRule(min, max));
-    });
-
-  g_lua.writeFunction("RCodeRule", [](int rcode) {
-      return std::shared_ptr<DNSRule>(new RCodeRule(rcode));
-    });
-
-  g_lua.writeFunction("addAction", [](luadnsrule_t var, boost::variant<std::shared_ptr<DNSAction>, std::shared_ptr<DNSResponseAction> > era)
-		      {
-                        if (era.type() == typeid(std::shared_ptr<DNSResponseAction>)) {
-                          throw std::runtime_error("addAction() can only be called with query-related actions, not response-related ones. Are you looking for addResponseAction()?");
-                        }
-
-                        auto ea = *boost::get<std::shared_ptr<DNSAction>>(&era);
-                        setLuaSideEffect();
-			auto rule=makeRule(var);
-			g_rulactions.modify([rule, ea](decltype(g_rulactions)::value_type& rulactions){
-			    rulactions.push_back({rule, ea});
-			  });
-		      });
-
-
-  g_lua.writeFunction("addPoolRule", [](luadnsrule_t var, string pool) {
-      setLuaSideEffect();
-      warnlog("addPoolRule() is deprecated and will be removed in 1.3.0, please use addAction(..., PoolAction(\"%s\")) instead", pool);
-
-      auto rule=makeRule(var);
-	g_rulactions.modify([rule, pool](decltype(g_rulactions)::value_type& rulactions) {
-	    rulactions.push_back({
-		rule,
-		  std::make_shared<PoolAction>(pool)  });
-	  });
-    });
-
-  g_lua.writeFunction("addNoRecurseRule", [](luadnsrule_t var) {
-      setLuaSideEffect();
-      warnlog("addNoRecurseRule() is deprecated and will be removed in 1.3.0, please use addAction(..., NoRecurseAction()) instead");
-
-      auto rule=makeRule(var);
-	g_rulactions.modify([rule](decltype(g_rulactions)::value_type& rulactions) {
-	    rulactions.push_back({
-		rule,
-		  std::make_shared<NoRecurseAction>()  });
-	  });
-    });
-
-  g_lua.writeFunction("addDisableValidationRule", [](luadnsrule_t var) {
-      setLuaSideEffect();
-      warnlog("addDisableValidationRule() is deprecated and will be removed in 1.3.0, please use addAction(..., DisableValidationAction()) instead");
-
-      auto rule=makeRule(var);
-	g_rulactions.modify([rule](decltype(g_rulactions)::value_type& rulactions) {
-	    rulactions.push_back({
-		rule,
-		  std::make_shared<DisableValidationAction>()  });
-	  });
-    });
-
-
-  g_lua.writeFunction("addQPSPoolRule", [](luadnsrule_t var, int limit, string pool) {
-      setLuaSideEffect();
-      warnlog("addQPSPoolRule() is deprecated and will be removed in 1.3.0, please use addAction(..., QPSPoolAction(%d, \"%s\")) instead", limit, pool);
-
-      auto rule = makeRule(var);
-      g_rulactions.modify([rule, pool,limit](decltype(g_rulactions)::value_type& rulactions) {
-	  rulactions.push_back({
-	      rule, 
-		std::make_shared<QPSPoolAction>(limit, pool)  });
-	});
-    });
-
-  g_lua.writeFunction("setDNSSECPool", [](const std::string& pool) {
-      setLuaSideEffect();
-      warnlog("setDNSSECPool() is deprecated and will be removed in 1.3.0, please use addAction(DNSSECRule(), PoolAction(\"%s\")) instead", pool);
-
-      g_rulactions.modify([pool](decltype(g_rulactions)::value_type& rulactions) {
-	  rulactions.push_back({std::make_shared<DNSSECRule>(), 
-		std::make_shared<PoolAction>(pool)}); 
-	});
-    });
-
-  g_lua.writeFunction("addQPSLimit", [](luadnsrule_t var, int lim) {
-      setLuaSideEffect();
-      warnlog("addQPSLimit() is deprecated and will be removed in 1.3.0, please use addAction(..., QPSAction(%d)) instead", lim);
-
-      auto rule = makeRule(var);
-      g_rulactions.modify([lim,rule](decltype(g_rulactions)::value_type& rulactions) {
-	  rulactions.push_back({rule, 
-		std::make_shared<QPSAction>(lim)});
-	});
-    });
-
-  g_lua.writeFunction("addDelay", [](luadnsrule_t var, int msec) {
-      setLuaSideEffect();
-      warnlog("addDelay() is deprecated and will be removed in 1.3.0, please use addAction(..., DelayAction(%d)) instead", msec);
-
-      auto rule = makeRule(var);
-      g_rulactions.modify([msec,rule](decltype(g_rulactions)::value_type& rulactions) {
-	  rulactions.push_back({rule, 
-		std::make_shared<DelayAction>(msec)});
-	});
-    });
-
-
-  g_lua.writeFunction("showRules", []() {
-     setLuaNoSideEffect();
-     boost::format fmt("%-3d %9d %-50s %s\n");
-     g_outputBuffer += (fmt % "#" % "Matches" % "Rule" % "Action").str();
-     int num=0;
-      for(const auto& lim : g_rulactions.getCopy()) {  
-        string name = lim.first->toString();
-	g_outputBuffer += (fmt % num % lim.first->d_matches % name % lim.second->toString()).str();
-	++num;
-      }
     });
 
   g_lua.writeFunction("getServers", []() {
@@ -1222,107 +550,6 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
         return std::make_shared<DownstreamState>(ComboAddress());
       return g_dstates.getCopy().at(i);
     });
-
-  g_lua.registerFunction<void(DownstreamState::*)(int)>("setQPS", [](DownstreamState& s, int lim) { s.qps = lim ? QPSLimiter(lim, lim) : QPSLimiter(); });
-  g_lua.registerFunction<void(std::shared_ptr<DownstreamState>::*)(string)>("addPool", [](std::shared_ptr<DownstreamState> s, string pool) {
-      auto localPools = g_pools.getCopy();
-      addServerToPool(localPools, pool, s);
-      g_pools.setState(localPools);
-      s->pools.insert(pool);
-    });
-  g_lua.registerFunction<void(std::shared_ptr<DownstreamState>::*)(string)>("rmPool", [](std::shared_ptr<DownstreamState> s, string pool) {
-      auto localPools = g_pools.getCopy();
-      removeServerFromPool(localPools, pool, s);
-      g_pools.setState(localPools);
-      s->pools.erase(pool);
-    });
-
-  g_lua.registerFunction<void(DownstreamState::*)()>("getOutstanding", [](const DownstreamState& s) { g_outputBuffer=std::to_string(s.outstanding.load()); });
-
-
-  g_lua.registerFunction("isUp", &DownstreamState::isUp);
-  g_lua.registerFunction("setDown", &DownstreamState::setDown);
-  g_lua.registerFunction("setUp", &DownstreamState::setUp);
-  g_lua.registerFunction<void(DownstreamState::*)(boost::optional<bool> newStatus)>("setAuto", [](DownstreamState& s, boost::optional<bool> newStatus) {
-      if (newStatus) {
-        s.upStatus = *newStatus;
-      }
-      s.setAuto();
-    });
-
-  g_lua.registerFunction("getName", &DownstreamState::getName);
-  g_lua.registerFunction("getNameWithAddr", &DownstreamState::getNameWithAddr);
-  g_lua.registerMember("upStatus", &DownstreamState::upStatus);
-  g_lua.registerMember("weight", &DownstreamState::weight);
-  g_lua.registerMember("order", &DownstreamState::order);
-  g_lua.registerMember("name", &DownstreamState::name);
-  
-  g_lua.writeFunction("infolog", [](const string& arg) {
-      infolog("%s", arg);
-    });
-  g_lua.writeFunction("errlog", [](const string& arg) {
-      errlog("%s", arg);
-    });
-  g_lua.writeFunction("warnlog", [](const string& arg) {
-      warnlog("%s", arg);
-    });
-
-
-  g_lua.writeFunction("show", [](const string& arg) {
-      g_outputBuffer+=arg;
-      g_outputBuffer+="\n";
-    });
-
-  g_lua.registerFunction<void(dnsheader::*)(bool)>("setRD", [](dnsheader& dh, bool v) {
-      dh.rd=v;
-    });
-
-  g_lua.registerFunction<bool(dnsheader::*)()>("getRD", [](dnsheader& dh) {
-      return (bool)dh.rd;
-    });
-
-  g_lua.registerFunction<void(dnsheader::*)(bool)>("setCD", [](dnsheader& dh, bool v) {
-      dh.cd=v;
-    });
-
-  g_lua.registerFunction<bool(dnsheader::*)()>("getCD", [](dnsheader& dh) {
-      return (bool)dh.cd;
-    });
-
-
-  g_lua.registerFunction<void(dnsheader::*)(bool)>("setTC", [](dnsheader& dh, bool v) {
-      dh.tc=v;
-      if(v) dh.ra = dh.rd; // you'll always need this, otherwise TC=1 gets ignored
-    });
-
-  g_lua.registerFunction<void(dnsheader::*)(bool)>("setQR", [](dnsheader& dh, bool v) {
-      dh.qr=v;
-    });
-
-  g_lua.registerFunction<string(std::shared_ptr<DNSRule>::*)()>("toString", [](const std::shared_ptr<DNSRule>& rule) { return rule->toString(); });
-
-  g_lua.registerFunction<string(ComboAddress::*)()>("tostring", [](const ComboAddress& ca) { return ca.toString(); });
-  g_lua.registerFunction<string(ComboAddress::*)()>("tostringWithPort", [](const ComboAddress& ca) { return ca.toStringWithPort(); });
-  g_lua.registerFunction<string(ComboAddress::*)()>("toString", [](const ComboAddress& ca) { return ca.toString(); });
-  g_lua.registerFunction<string(ComboAddress::*)()>("toStringWithPort", [](const ComboAddress& ca) { return ca.toStringWithPort(); });
-  g_lua.registerFunction<uint16_t(ComboAddress::*)()>("getPort", [](const ComboAddress& ca) { return ntohs(ca.sin4.sin_port); } );
-  g_lua.registerFunction<void(ComboAddress::*)(unsigned int)>("truncate", [](ComboAddress& ca, unsigned int bits) { ca.truncate(bits); });
-  g_lua.registerFunction<bool(ComboAddress::*)()>("isIPv4", [](const ComboAddress& ca) { return ca.sin4.sin_family == AF_INET; });
-  g_lua.registerFunction<bool(ComboAddress::*)()>("isIPv6", [](const ComboAddress& ca) { return ca.sin4.sin_family == AF_INET6; });
-  g_lua.registerFunction<bool(ComboAddress::*)()>("isMappedIPv4", [](const ComboAddress& ca) { return ca.isMappedIPv4(); });
-  g_lua.registerFunction<ComboAddress(ComboAddress::*)()>("mapToIPv4", [](const ComboAddress& ca) { return ca.mapToIPv4(); });
-
-  g_lua.registerFunction("isPartOf", &DNSName::isPartOf);
-  g_lua.registerFunction<bool(DNSName::*)()>("chopOff", [](DNSName&dn ) { return dn.chopOff(); });
-  g_lua.registerFunction<unsigned int(DNSName::*)()>("countLabels", [](const DNSName& name) { return name.countLabels(); });
-  g_lua.registerFunction<size_t(DNSName::*)()>("wirelength", [](const DNSName& name) { return name.wirelength(); });
-  g_lua.registerFunction<string(DNSName::*)()>("tostring", [](const DNSName&dn ) { return dn.toString(); });
-  g_lua.registerFunction<string(DNSName::*)()>("toString", [](const DNSName&dn ) { return dn.toString(); });
-  g_lua.writeFunction("newDNSName", [](const std::string& name) { return DNSName(name); });
-  g_lua.writeFunction("newSuffixMatchNode", []() { return SuffixMatchNode(); });
-
-  g_lua.registerFunction("add",(void (SuffixMatchNode::*)(const DNSName&)) &SuffixMatchNode::add);
-  g_lua.registerFunction("check",(bool (SuffixMatchNode::*)(const DNSName&) const) &SuffixMatchNode::check);
 
   g_lua.writeFunction("carbonServer", [](const std::string& address, boost::optional<string> ourName,
 					 boost::optional<unsigned int> interval) {
@@ -1346,10 +573,10 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
 	  thread t(dnsdistWebserverThread, sock, local, password, apiKey ? *apiKey : "", customHeaders);
 	  t.detach();
 	};
-	if(g_launchWork) 
+	if(g_launchWork)
 	  g_launchWork->push_back(launch);
 	else
-	  launch();	    
+	  launch();
       }
       catch(std::exception& e) {
 	g_outputBuffer="Unable to bind to webserver socket on " + local.toStringWithPort() + ": " + e.what();
@@ -1357,6 +584,7 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
       }
 
     });
+
   g_lua.writeFunction("controlSocket", [client](const std::string& str) {
       setLuaSideEffect();
       ComboAddress local(str, 5199);
@@ -1365,7 +593,7 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
 	g_serverControl = local;
 	return;
       }
-      
+
       try {
 	int sock = SSocket(local.sin4.sin_family, SOCK_STREAM, 0);
 	SSetsockopt(sock, SOL_SOCKET, SO_REUSEADDR, 1);
@@ -1375,96 +603,17 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
 	    thread t(controlThread, sock, local);
 	    t.detach();
 	};
-	if(g_launchWork) 
+	if(g_launchWork)
 	  g_launchWork->push_back(launch);
 	else
 	  launch();
-	    
+
       }
       catch(std::exception& e) {
 	g_outputBuffer="Unable to bind to control socket on " + local.toStringWithPort() + ": " + e.what();
 	errlog("Unable to bind to control socket on %s: %s", local.toStringWithPort(), e.what());
       }
     });
-
-
-  g_lua.writeFunction("topClients", [](boost::optional<unsigned int> top_) {
-      setLuaNoSideEffect();
-      auto top = top_.get_value_or(10);
-      map<ComboAddress, int,ComboAddress::addressOnlyLessThan > counts;
-      unsigned int total=0;
-      {
-        ReadLock rl(&g_rings.queryLock);
-        for(const auto& c : g_rings.queryRing) {
-          counts[c.requestor]++;
-          total++;
-        }
-      }
-      vector<pair<int, ComboAddress>> rcounts;
-      rcounts.reserve(counts.size());
-      for(const auto& c : counts) 
-	rcounts.push_back(make_pair(c.second, c.first));
-
-      sort(rcounts.begin(), rcounts.end(), [](const decltype(rcounts)::value_type& a, 
-					      const decltype(rcounts)::value_type& b) {
-	     return b.first < a.first;
-	   });
-      unsigned int count=1, rest=0;
-      boost::format fmt("%4d  %-40s %4d %4.1f%%\n");
-      for(const auto& rc : rcounts) {
-	if(count==top+1)
-	  rest+=rc.first;
-	else
-	  g_outputBuffer += (fmt % (count++) % rc.second.toString() % rc.first % (100.0*rc.first/total)).str();
-      }
-      g_outputBuffer += (fmt % (count) % "Rest" % rest % (total > 0 ? 100.0*rest/total : 100.0)).str();
-    });
-
-  g_lua.writeFunction("getTopQueries", [](unsigned int top, boost::optional<int> labels) {
-      setLuaNoSideEffect();
-      map<DNSName, int> counts;
-      unsigned int total=0;
-      if(!labels) {
-	ReadLock rl(&g_rings.queryLock);
-	for(const auto& a : g_rings.queryRing) {
-	  counts[a.name]++;
-	  total++;
-	}
-      }
-      else {
-	unsigned int lab = *labels;
-	ReadLock rl(&g_rings.queryLock);
-	for(auto a : g_rings.queryRing) {
-	  a.name.trimToLabels(lab);
-	  counts[a.name]++;
-	  total++;
-	}
-      }
-      // cout<<"Looked at "<<total<<" queries, "<<counts.size()<<" different ones"<<endl;
-      vector<pair<int, DNSName>> rcounts;
-      rcounts.reserve(counts.size());
-      for(const auto& c : counts) 
-	rcounts.push_back(make_pair(c.second, c.first.makeLowerCase()));
-
-      sort(rcounts.begin(), rcounts.end(), [](const decltype(rcounts)::value_type& a, 
-					      const decltype(rcounts)::value_type& b) {
-	     return b.first < a.first;
-	   });
-
-      std::unordered_map<int, vector<boost::variant<string,double>>> ret;
-      unsigned int count=1, rest=0;
-      for(const auto& rc : rcounts) {
-	if(count==top+1)
-	  rest+=rc.first;
-	else
-	  ret.insert({count++, {rc.second.toString(), rc.first, 100.0*rc.first/total}});
-      }
-      ret.insert({count, {"Rest", rest, total > 0 ? 100.0*rest/total : 100.0}});
-      return ret;
-
-    });
-
-  g_lua.executeCode(R"(function topQueries(top, labels) top = top or 10; for k,v in ipairs(getTopQueries(top,labels)) do show(string.format("%4d  %-40s %4d %4.1f%%",k,v[1],v[2], v[3])) end end)");
 
   g_lua.writeFunction("clearQueryCounters", []() {
       unsigned int size{0};
@@ -1495,110 +644,16 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
     });
 
   g_lua.writeFunction("setQueryCount", [](bool enabled) { g_qcount.enabled=enabled; });
+
   g_lua.writeFunction("setQueryCountFilter", [](QueryCountFilter func) {
       g_qcount.filter = func;
     });
-
-  g_lua.writeFunction("getResponseRing", []() {
-      setLuaNoSideEffect();
-      decltype(g_rings.respRing) ring;
-      {
-	std::lock_guard<std::mutex> lock(g_rings.respMutex);
-	ring = g_rings.respRing;
-      }
-      vector<std::unordered_map<string, boost::variant<string, unsigned int> > > ret;
-      ret.reserve(ring.size());
-      decltype(ret)::value_type item;
-      for(const auto& r : ring) {
-	item["name"]=r.name.toString();
-	item["qtype"]=r.qtype;
-	item["rcode"]=r.dh.rcode;
-	item["usec"]=r.usec;
-	ret.push_back(item);
-      }
-      return ret;
-    });
-
-  g_lua.writeFunction("getTopResponses", [](unsigned int top, unsigned int kind, boost::optional<int> labels) {
-      return getGenResponses(top, labels, [kind](const Rings::Response& r) { return r.dh.rcode == kind; });
-    });
-
-  g_lua.executeCode(R"(function topResponses(top, kind, labels) top = top or 10; kind = kind or 0; for k,v in ipairs(getTopResponses(top, kind, labels)) do show(string.format("%4d  %-40s %4d %4.1f%%",k,v[1],v[2],v[3])) end end)");
-
-
-  g_lua.writeFunction("getSlowResponses", [](unsigned int top, unsigned int msec, boost::optional<int> labels) {
-      return getGenResponses(top, labels, [msec](const Rings::Response& r) { return r.usec > msec*1000; });
-    });
-
-
-  g_lua.executeCode(R"(function topSlow(top, msec, labels) top = top or 10; msec = msec or 500; for k,v in ipairs(getSlowResponses(top, msec, labels)) do show(string.format("%4d  %-40s %4d %4.1f%%",k,v[1],v[2],v[3])) end end)");
-
-
-  g_lua.writeFunction("showResponseLatency", []() {
-      setLuaNoSideEffect();
-      map<double, unsigned int> histo;
-      double bin=100;
-      for(int i=0; i < 15; ++i) {
-	histo[bin];
-	bin*=2;
-      }
-
-      double totlat=0;
-      unsigned int size=0;
-      {
-	std::lock_guard<std::mutex> lock(g_rings.respMutex);
-	for(const auto& r : g_rings.respRing) {
-          /* skip actively discovered timeouts */
-          if (r.usec == std::numeric_limits<unsigned int>::max())
-            continue;
-
-	  ++size;
-	  auto iter = histo.lower_bound(r.usec);
-	  if(iter != histo.end())
-	    iter->second++;
-	  else
-	    histo.rbegin()++;
-	  totlat+=r.usec;
-	}
-      }
-
-      if (size == 0) {
-        g_outputBuffer = "No traffic yet.\n";
-        return;
-      }
-
-      g_outputBuffer = (boost::format("Average response latency: %.02f msec\n") % (0.001*totlat/size)).str();
-      double highest=0;
-      
-      for(auto iter = histo.cbegin(); iter != histo.cend(); ++iter) {
-	highest=std::max(highest, iter->second*1.0);
-      }
-      boost::format fmt("%7.2f\t%s\n");
-      g_outputBuffer += (fmt % "msec" % "").str();
-
-      for(auto iter = histo.cbegin(); iter != histo.cend(); ++iter) {
-	int stars = (70.0 * iter->second/highest);
-	char c='*';
-	if(!stars && iter->second) {
-	  stars=1; // you get 1 . to show something is there..
-	  if(70.0*iter->second/highest > 0.5)
-	    c=':';
-	  else
-	    c='.';
-	}
-	g_outputBuffer += (fmt % (iter->first/1000.0) % string(stars, c)).str();
-      }
-    });
-
-  g_lua.writeFunction("newQPSLimiter", [](int rate, int burst) { return QPSLimiter(rate, burst); });
-  g_lua.registerFunction("check", &QPSLimiter::check);
-
 
   g_lua.writeFunction("makeKey", []() {
       setLuaNoSideEffect();
       g_outputBuffer="setKey("+newKey()+")\n";
     });
-  
+
   g_lua.writeFunction("setKey", [](const std::string& key) {
       if(!g_configurationDone && ! g_key.empty()) { // this makes sure the commandline -k key prevails over dnsdist.conf
         return;                                     // but later setKeys() trump the -k value again
@@ -1614,7 +669,6 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
 	g_key=newkey;
     });
 
-  
   g_lua.writeFunction("testCrypto", [](boost::optional<string> optTestMsg)
    {
      setLuaNoSideEffect();
@@ -1634,7 +688,7 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
        sn2=sn;
        string encrypted = sodEncryptSym(testmsg, g_key, sn);
        string decrypted = sodDecryptSym(encrypted, g_key, sn2);
-       
+
        sn.increment();
        sn2.increment();
 
@@ -1645,7 +699,7 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
 	 g_outputBuffer="Everything is ok!\n";
        else
 	 g_outputBuffer="Crypto failed..\n";
-       
+
      }
      catch(...) {
        g_outputBuffer="Crypto failed..\n";
@@ -1667,102 +721,6 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
       } else {
         g_outputBuffer="Max UDP outstanding cannot be altered at runtime!\n";
       }
-    });
-
-  g_lua.registerFunction<void(DNSQuestion::*)(std::string, std::string)>("setTag", [](DNSQuestion& dq, const std::string& strLabel, const std::string& strValue) {
-
-      if(dq.qTag == nullptr) {
-        dq.qTag = std::make_shared<QTag>();
-      }
-      dq.qTag->add(strLabel, strValue);
-
-    });
-
-  g_lua.registerFunction<void(DNSQuestion::*)(vector<pair<string, string>>)>("setTagArray", [](DNSQuestion& dq, const vector<pair<string, string>>&tags) {
-
-      if(dq.qTag == nullptr) {
-        dq.qTag = std::make_shared<QTag>();
-      }
-
-      for (const auto& tag : tags) {
-        dq.qTag->add(tag.first, tag.second);
-      }
-
-    });
-
-  g_lua.registerFunction<string(DNSQuestion::*)(std::string)>("getTag", [](const DNSQuestion& dq, const std::string& strLabel) {
-
-      std::string strValue;
-      if(dq.qTag != nullptr) {
-        strValue = dq.qTag->getMatch(strLabel);
-      }
-      return strValue;
-
-    });
-
-
-  g_lua.registerFunction<std::unordered_map<string, string>(DNSQuestion::*)(void)>("getTagArray", [](const DNSQuestion& dq) {
-
-      if(dq.qTag != nullptr) {
-        return dq.qTag->tagData;
-      } else {
-        std::unordered_map<string, string> XX;
-        return XX;
-      }
-    });
-
-
-  /* DNSQuestion bindings */
-  /* PowerDNS DNSQuestion compat */
-  g_lua.registerMember<const ComboAddress (DNSQuestion::*)>("localaddr", [](const DNSQuestion& dq) -> const ComboAddress { return *dq.local; }, [](DNSQuestion& dq, const ComboAddress newLocal) { (void) newLocal; });
-  g_lua.registerMember<const DNSName (DNSQuestion::*)>("qname", [](const DNSQuestion& dq) -> const DNSName { return *dq.qname; }, [](DNSQuestion& dq, const DNSName newName) { (void) newName; });
-  g_lua.registerMember<uint16_t (DNSQuestion::*)>("qtype", [](const DNSQuestion& dq) -> uint16_t { return dq.qtype; }, [](DNSQuestion& dq, uint16_t newType) { (void) newType; });
-  g_lua.registerMember<uint16_t (DNSQuestion::*)>("qclass", [](const DNSQuestion& dq) -> uint16_t { return dq.qclass; }, [](DNSQuestion& dq, uint16_t newClass) { (void) newClass; });
-  g_lua.registerMember<int (DNSQuestion::*)>("rcode", [](const DNSQuestion& dq) -> int { return dq.dh->rcode; }, [](DNSQuestion& dq, int newRCode) { dq.dh->rcode = newRCode; });
-  g_lua.registerMember<const ComboAddress (DNSQuestion::*)>("remoteaddr", [](const DNSQuestion& dq) -> const ComboAddress { return *dq.remote; }, [](DNSQuestion& dq, const ComboAddress newRemote) { (void) newRemote; });
-  /* DNSDist DNSQuestion */
-  g_lua.registerMember("dh", &DNSQuestion::dh);
-  g_lua.registerMember<uint16_t (DNSQuestion::*)>("len", [](const DNSQuestion& dq) -> uint16_t { return dq.len; }, [](DNSQuestion& dq, uint16_t newlen) { dq.len = newlen; });
-  g_lua.registerMember<uint8_t (DNSQuestion::*)>("opcode", [](const DNSQuestion& dq) -> uint8_t { return dq.dh->opcode; }, [](DNSQuestion& dq, uint8_t newOpcode) { (void) newOpcode; });
-  g_lua.registerMember<size_t (DNSQuestion::*)>("size", [](const DNSQuestion& dq) -> size_t { return dq.size; }, [](DNSQuestion& dq, size_t newSize) { (void) newSize; });
-  g_lua.registerMember<bool (DNSQuestion::*)>("tcp", [](const DNSQuestion& dq) -> bool { return dq.tcp; }, [](DNSQuestion& dq, bool newTcp) { (void) newTcp; });
-  g_lua.registerMember<bool (DNSQuestion::*)>("skipCache", [](const DNSQuestion& dq) -> bool { return dq.skipCache; }, [](DNSQuestion& dq, bool newSkipCache) { dq.skipCache = newSkipCache; });
-  g_lua.registerMember<bool (DNSQuestion::*)>("useECS", [](const DNSQuestion& dq) -> bool { return dq.useECS; }, [](DNSQuestion& dq, bool useECS) { dq.useECS = useECS; });
-  g_lua.registerMember<bool (DNSQuestion::*)>("ecsOverride", [](const DNSQuestion& dq) -> bool { return dq.ecsOverride; }, [](DNSQuestion& dq, bool ecsOverride) { dq.ecsOverride = ecsOverride; });
-  g_lua.registerMember<uint16_t (DNSQuestion::*)>("ecsPrefixLength", [](const DNSQuestion& dq) -> uint16_t { return dq.ecsPrefixLength; }, [](DNSQuestion& dq, uint16_t newPrefixLength) { dq.ecsPrefixLength = newPrefixLength; });
-  g_lua.registerFunction<bool(DNSQuestion::*)()>("getDO", [](const DNSQuestion& dq) {
-      return getEDNSZ((const char*)dq.dh, dq.len) & EDNS_HEADER_FLAG_DO;
-    });
-  g_lua.registerFunction<void(DNSQuestion::*)(std::string)>("sendTrap", [](const DNSQuestion& dq, boost::optional<std::string> reason) {
-#ifdef HAVE_NET_SNMP
-      if (g_snmpAgent && g_snmpTrapsEnabled) {
-        g_snmpAgent->sendDNSTrap(dq, reason ? *reason : "");
-      }
-#endif /* HAVE_NET_SNMP */
-    });
-
-  /* LuaWrapper doesn't support inheritance */
-  g_lua.registerMember<const ComboAddress (DNSResponse::*)>("localaddr", [](const DNSResponse& dq) -> const ComboAddress { return *dq.local; }, [](DNSResponse& dq, const ComboAddress newLocal) { (void) newLocal; });
-  g_lua.registerMember<const DNSName (DNSResponse::*)>("qname", [](const DNSResponse& dq) -> const DNSName { return *dq.qname; }, [](DNSResponse& dq, const DNSName newName) { (void) newName; });
-  g_lua.registerMember<uint16_t (DNSResponse::*)>("qtype", [](const DNSResponse& dq) -> uint16_t { return dq.qtype; }, [](DNSResponse& dq, uint16_t newType) { (void) newType; });
-  g_lua.registerMember<uint16_t (DNSResponse::*)>("qclass", [](const DNSResponse& dq) -> uint16_t { return dq.qclass; }, [](DNSResponse& dq, uint16_t newClass) { (void) newClass; });
-  g_lua.registerMember<int (DNSResponse::*)>("rcode", [](const DNSResponse& dq) -> int { return dq.dh->rcode; }, [](DNSResponse& dq, int newRCode) { dq.dh->rcode = newRCode; });
-  g_lua.registerMember<const ComboAddress (DNSResponse::*)>("remoteaddr", [](const DNSResponse& dq) -> const ComboAddress { return *dq.remote; }, [](DNSResponse& dq, const ComboAddress newRemote) { (void) newRemote; });
-  g_lua.registerMember<dnsheader* (DNSResponse::*)>("dh", [](const DNSResponse& dr) -> dnsheader* { return dr.dh; }, [](DNSResponse& dr, dnsheader * newdh) { dr.dh = newdh; });
-  g_lua.registerMember<uint16_t (DNSResponse::*)>("len", [](const DNSResponse& dq) -> uint16_t { return dq.len; }, [](DNSResponse& dq, uint16_t newlen) { dq.len = newlen; });
-  g_lua.registerMember<uint8_t (DNSResponse::*)>("opcode", [](const DNSResponse& dq) -> uint8_t { return dq.dh->opcode; }, [](DNSResponse& dq, uint8_t newOpcode) { (void) newOpcode; });
-  g_lua.registerMember<size_t (DNSResponse::*)>("size", [](const DNSResponse& dq) -> size_t { return dq.size; }, [](DNSResponse& dq, size_t newSize) { (void) newSize; });
-  g_lua.registerMember<bool (DNSResponse::*)>("tcp", [](const DNSResponse& dq) -> bool { return dq.tcp; }, [](DNSResponse& dq, bool newTcp) { (void) newTcp; });
-  g_lua.registerMember<bool (DNSResponse::*)>("skipCache", [](const DNSResponse& dq) -> bool { return dq.skipCache; }, [](DNSResponse& dq, bool newSkipCache) { dq.skipCache = newSkipCache; });
-  g_lua.registerFunction<void(DNSResponse::*)(std::function<uint32_t(uint8_t section, uint16_t qclass, uint16_t qtype, uint32_t ttl)> editFunc)>("editTTLs", [](const DNSResponse& dr, std::function<uint32_t(uint8_t section, uint16_t qclass, uint16_t qtype, uint32_t ttl)> editFunc) {
-        editDNSPacketTTL((char*) dr.dh, dr.len, editFunc);
-      });
-  g_lua.registerFunction<void(DNSResponse::*)(std::string)>("sendTrap", [](const DNSResponse& dr, boost::optional<std::string> reason) {
-#ifdef HAVE_NET_SNMP
-      if (g_snmpAgent && g_snmpTrapsEnabled) {
-        g_snmpAgent->sendDNSTrap(dr, reason ? *reason : "");
-      }
-#endif /* HAVE_NET_SNMP */
     });
 
   g_lua.writeFunction("setMaxTCPClientThreads", [](uint64_t max) {
@@ -1805,15 +763,8 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
       }
     });
 
-  g_lua.writeFunction("showTCPStats", [] {
-      setLuaNoSideEffect();
-      boost::format fmt("%-10d %-10d %-10d %-10d\n");
-      g_outputBuffer += (fmt % "Clients" % "MaxClients" % "Queued" % "MaxQueued").str();
-      g_outputBuffer += (fmt % g_tcpclientthreads->getThreadsCount() % g_maxTCPClientThreads % g_tcpclientthreads->getQueuedCount() % g_maxTCPQueuedConnections).str();
-      g_outputBuffer += "Query distribution mode is: " + std::string(g_useTCPSinglePipe ? "single queue" : "per-thread queues") + "\n";
-    });
-
   g_lua.writeFunction("setCacheCleaningDelay", [](uint32_t delay) { g_cacheCleaningDelay = delay; });
+
   g_lua.writeFunction("setCacheCleaningPercentage", [](uint16_t percentage) { if (percentage < 100) g_cacheCleaningPercentage = percentage; else g_cacheCleaningPercentage = 100; });
 
   g_lua.writeFunction("setECSSourcePrefixV4", [](uint16_t prefix) { g_ECSSourcePrefixV4=prefix; });
@@ -1822,62 +773,618 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
 
   g_lua.writeFunction("setECSOverride", [](bool override) { g_ECSOverride=override; });
 
-  g_lua.writeFunction("dumpStats", [] {
+  g_lua.writeFunction("showDynBlocks", []() {
       setLuaNoSideEffect();
-      vector<string> leftcolumn, rightcolumn;
-
-      boost::format fmt("%-23s\t%+11s");
-      g_outputBuffer.clear();
-      auto entries = g_stats.entries;
-      sort(entries.begin(), entries.end(), 
-	   [](const decltype(entries)::value_type& a, const decltype(entries)::value_type& b) {
-	     return a.first < b.first;
-	   });
-      boost::format flt("    %9.1f");
-      for(const auto& e : entries) {
-	string second;
-	if(const auto& val = boost::get<DNSDistStats::stat_t*>(&e.second))
-	  second=std::to_string((*val)->load());
-	else if (const auto& dval = boost::get<double*>(&e.second))
-	  second=(flt % (**dval)).str();
-	else
-	  second=std::to_string((*boost::get<DNSDistStats::statfunction_t>(&e.second))(e.first));
-
-	if(leftcolumn.size() < g_stats.entries.size()/2)
-	  leftcolumn.push_back((fmt % e.first % second).str());
-	else
-	  rightcolumn.push_back((fmt % e.first % second).str());
+      auto slow = g_dynblockNMG.getCopy();
+      struct timespec now;
+      gettime(&now);
+      boost::format fmt("%-24s %8d %8d %s\n");
+      g_outputBuffer = (fmt % "What" % "Seconds" % "Blocks" % "Reason").str();
+      for(const auto& e: slow) {
+	if(now < e->second.until)
+	  g_outputBuffer+= (fmt % e->first.toString() % (e->second.until.tv_sec - now.tv_sec) % e->second.blocks % e->second.reason).str();
       }
+      auto slow2 = g_dynblockSMT.getCopy();
+      slow2.visit([&now, &fmt](const SuffixMatchTree<DynBlock>& node) {
+          if(now <node.d_value.until) {
+            string dom("empty");
+            if(!node.d_value.domain.empty())
+              dom = node.d_value.domain.toString();
+            g_outputBuffer+= (fmt % dom % (node.d_value.until.tv_sec - now.tv_sec) % node.d_value.blocks % node.d_value.reason).str();
+          }
+        });
 
-      auto leftiter=leftcolumn.begin(), rightiter=rightcolumn.begin();
-      boost::format clmn("%|0t|%1% %|39t|%2%\n");
+    });
 
-      for(;leftiter != leftcolumn.end() || rightiter != rightcolumn.end();) {
-	string lentry, rentry;
-	if(leftiter!= leftcolumn.end()) {
-	  lentry = *leftiter;
-	  leftiter++;
-	}
-	if(rightiter!= rightcolumn.end()) {
-	  rentry = *rightiter;
-	  rightiter++;
-	}
-	g_outputBuffer += (clmn % lentry % rentry).str();
+  g_lua.writeFunction("clearDynBlocks", []() {
+      setLuaSideEffect();
+      nmts_t nmg;
+      g_dynblockNMG.setState(nmg);
+      SuffixMatchTree<DynBlock> smt;
+      g_dynblockSMT.setState(smt);
+    });
+
+  g_lua.writeFunction("addDynBlocks",
+                      [](const map<ComboAddress,int>& m, const std::string& msg, boost::optional<int> seconds, boost::optional<DNSAction::Action> action) {
+                           setLuaSideEffect();
+			   auto slow = g_dynblockNMG.getCopy();
+			   struct timespec until, now;
+			   gettime(&now);
+			   until=now;
+                           int actualSeconds = seconds ? *seconds : 10;
+			   until.tv_sec += actualSeconds;
+			   for(const auto& capair : m) {
+			     unsigned int count = 0;
+                             auto got = slow.lookup(Netmask(capair.first));
+                             bool expired=false;
+			     if(got) {
+			       if(until < got->second.until) // had a longer policy
+				 continue;
+			       if(now < got->second.until) // only inherit count on fresh query we are extending
+				 count=got->second.blocks;
+                               else
+                                 expired=true;
+			     }
+			     DynBlock db{msg,until,DNSName(),(action ? *action : DNSAction::Action::None)};
+			     db.blocks=count;
+                             if(!got || expired)
+                               warnlog("Inserting dynamic block for %s for %d seconds: %s", capair.first.toString(), actualSeconds, msg);
+			     slow.insert(Netmask(capair.first)).second=db;
+			   }
+			   g_dynblockNMG.setState(slow);
+			 });
+
+  g_lua.writeFunction("addDynBlockSMT",
+                      [](const vector<pair<unsigned int, string> >&names, const std::string& msg, boost::optional<int> seconds, boost::optional<DNSAction::Action> action) {
+                           setLuaSideEffect();
+			   auto slow = g_dynblockSMT.getCopy();
+			   struct timespec until, now;
+			   gettime(&now);
+			   until=now;
+                           int actualSeconds = seconds ? *seconds : 10;
+			   until.tv_sec += actualSeconds;
+
+			   for(const auto& capair : names) {
+			     unsigned int count = 0;
+                             DNSName domain(capair.second);
+                             auto got = slow.lookup(domain);
+                             bool expired=false;
+			     if(got) {
+			       if(until < got->until) // had a longer policy
+				 continue;
+			       if(now < got->until) // only inherit count on fresh query we are extending
+				 count=got->blocks;
+                               else
+                                 expired=true;
+			     }
+
+			     DynBlock db{msg,until,domain,(action ? *action : DNSAction::Action::None)};
+			     db.blocks=count;
+                             if(!got || expired)
+                               warnlog("Inserting dynamic block for %s for %d seconds: %s", domain, actualSeconds, msg);
+			     slow.add(domain, db);
+			   }
+			   g_dynblockSMT.setState(slow);
+			 });
+
+  g_lua.writeFunction("setDynBlocksAction", [](DNSAction::Action action) {
+      if (!g_configurationDone) {
+        if (action == DNSAction::Action::Drop || action == DNSAction::Action::Refused || action == DNSAction::Action::Truncate) {
+          g_dynBlockAction = action;
+        }
+        else {
+          errlog("Dynamic blocks action can only be Drop, Refused or Truncate!");
+          g_outputBuffer="Dynamic blocks action can only be Drop, Refused or Truncate!\n";
+        }
+      } else {
+        g_outputBuffer="Dynamic blocks action cannot be altered at runtime!\n";
       }
     });
 
-  moreLua(client);
-  
+  g_lua.writeFunction("addDNSCryptBind", [](const std::string& addr, const std::string& providerName, const std::string& certFile, const std::string keyFile, boost::optional<localbind_t> vars) {
+      if (g_configurationDone) {
+        g_outputBuffer="addDNSCryptBind cannot be used at runtime!\n";
+        return;
+      }
+#ifdef HAVE_DNSCRYPT
+      bool doTCP = true;
+      bool reusePort = false;
+      int tcpFastOpenQueueSize = 0;
+      std::string interface;
+      std::set<int> cpus;
+
+      parseLocalBindVars(vars, doTCP, reusePort, tcpFastOpenQueueSize, interface, cpus);
+
+      try {
+        DnsCryptContext ctx(providerName, certFile, keyFile);
+        g_dnsCryptLocals.push_back(std::make_tuple(ComboAddress(addr, 443), ctx, reusePort, tcpFastOpenQueueSize, interface, cpus));
+      }
+      catch(std::exception& e) {
+        errlog(e.what());
+	g_outputBuffer="Error: "+string(e.what())+"\n";
+      }
+#else
+      g_outputBuffer="Error: DNSCrypt support is not enabled.\n";
+#endif
+    });
+
+  g_lua.writeFunction("showDNSCryptBinds", []() {
+      setLuaNoSideEffect();
+#ifdef HAVE_DNSCRYPT
+      ostringstream ret;
+      boost::format fmt("%1$-3d %2% %|25t|%3$-20.20s %|26t|%4$-8d %|35t|%5$-21.21s %|56t|%6$-9d %|66t|%7$-21.21s" );
+      ret << (fmt % "#" % "Address" % "Provider Name" % "Serial" % "Validity" % "P. Serial" % "P. Validity") << endl;
+      size_t idx = 0;
+
+      for (const auto& local : g_dnsCryptLocals) {
+        const DnsCryptContext& ctx = std::get<1>(local);
+        bool const hasOldCert = ctx.hasOldCertificate();
+        const DnsCryptCert& cert = ctx.getCurrentCertificate();
+        const DnsCryptCert& oldCert = ctx.getOldCertificate();
+
+        ret<< (fmt % idx % std::get<0>(local).toStringWithPort() % ctx.getProviderName() % cert.signedData.serial % DnsCryptContext::certificateDateToStr(cert.signedData.tsEnd) % (hasOldCert ? oldCert.signedData.serial : 0) % (hasOldCert ? DnsCryptContext::certificateDateToStr(oldCert.signedData.tsEnd) : "-")) << endl;
+        idx++;
+      }
+
+      g_outputBuffer=ret.str();
+#else
+      g_outputBuffer="Error: DNSCrypt support is not enabled.\n";
+#endif
+    });
+
+  g_lua.writeFunction("getDNSCryptBind", [client](size_t idx) {
+      setLuaNoSideEffect();
+#ifdef HAVE_DNSCRYPT
+      DnsCryptContext* ret = nullptr;
+      if (idx < g_dnsCryptLocals.size()) {
+        ret = &(std::get<1>(g_dnsCryptLocals.at(idx)));
+      }
+      return ret;
+#else
+      g_outputBuffer="Error: DNSCrypt support is not enabled.\n";
+#endif
+    });
+
+  g_lua.writeFunction("generateDNSCryptProviderKeys", [](const std::string& publicKeyFile, const std::string privateKeyFile) {
+      setLuaNoSideEffect();
+#ifdef HAVE_DNSCRYPT
+      unsigned char publicKey[DNSCRYPT_PROVIDER_PUBLIC_KEY_SIZE];
+      unsigned char privateKey[DNSCRYPT_PROVIDER_PRIVATE_KEY_SIZE];
+      sodium_mlock(privateKey, sizeof(privateKey));
+
+      try {
+        DnsCryptContext::generateProviderKeys(publicKey, privateKey);
+
+        ofstream pubKStream(publicKeyFile);
+        pubKStream.write((char*) publicKey, sizeof(publicKey));
+        pubKStream.close();
+
+        ofstream privKStream(privateKeyFile);
+        privKStream.write((char*) privateKey, sizeof(privateKey));
+        privKStream.close();
+
+        g_outputBuffer="Provider fingerprint is: " + DnsCryptContext::getProviderFingerprint(publicKey) + "\n";
+      }
+      catch(std::exception& e) {
+        errlog(e.what());
+        g_outputBuffer="Error: "+string(e.what())+"\n";
+      }
+
+      sodium_memzero(privateKey, sizeof(privateKey));
+      sodium_munlock(privateKey, sizeof(privateKey));
+#else
+      g_outputBuffer="Error: DNSCrypt support is not enabled.\n";
+#endif
+    });
+
+  g_lua.writeFunction("printDNSCryptProviderFingerprint", [](const std::string& publicKeyFile) {
+      setLuaNoSideEffect();
+#ifdef HAVE_DNSCRYPT
+      unsigned char publicKey[DNSCRYPT_PROVIDER_PUBLIC_KEY_SIZE];
+
+      try {
+        ifstream file(publicKeyFile);
+        file.read((char *) &publicKey, sizeof(publicKey));
+
+        if (file.fail())
+          throw std::runtime_error("Invalid dnscrypt provider public key file " + publicKeyFile);
+
+        file.close();
+        g_outputBuffer="Provider fingerprint is: " + DnsCryptContext::getProviderFingerprint(publicKey) + "\n";
+      }
+      catch(std::exception& e) {
+        errlog(e.what());
+        g_outputBuffer="Error: "+string(e.what())+"\n";
+      }
+#else
+      g_outputBuffer="Error: DNSCrypt support is not enabled.\n";
+#endif
+    });
+
+  g_lua.writeFunction("generateDNSCryptCertificate", [](const std::string& providerPrivateKeyFile, const std::string& certificateFile, const std::string privateKeyFile, uint32_t serial, time_t begin, time_t end) {
+      setLuaNoSideEffect();
+#ifdef HAVE_DNSCRYPT
+      DnsCryptPrivateKey privateKey;
+      DnsCryptCert cert;
+
+      try {
+        if (generateDNSCryptCertificate(providerPrivateKeyFile, serial, begin, end, cert, privateKey)) {
+          privateKey.saveToFile(privateKeyFile);
+          DnsCryptContext::saveCertFromFile(cert, certificateFile);
+        }
+      }
+      catch(const std::exception& e) {
+        errlog(e.what());
+        g_outputBuffer="Error: "+string(e.what())+"\n";
+      }
+#else
+      g_outputBuffer="Error: DNSCrypt support is not enabled.\n";
+#endif
+    });
+
+  g_lua.writeFunction("showPools", []() {
+      setLuaNoSideEffect();
+      try {
+        ostringstream ret;
+        boost::format fmt("%1$-20.20s %|25t|%2$20s %|25t|%3$20s %|50t|%4%" );
+        //             1        2         3                4
+        ret << (fmt % "Name" % "Cache" % "ServerPolicy" % "Servers" ) << endl;
+
+        const auto localPools = g_pools.getCopy();
+        for (const auto& entry : localPools) {
+          const string& name = entry.first;
+          const std::shared_ptr<ServerPool> pool = entry.second;
+          string cache = pool->packetCache != nullptr ? pool->packetCache->toString() : "";
+          string policy = g_policy.getLocal()->name;
+          if (pool->policy != nullptr) {
+            policy = pool->policy->name;
+          }
+          string servers;
+
+          for (const auto& server: pool->servers) {
+            if (!servers.empty()) {
+              servers += ", ";
+            }
+            if (!server.second->name.empty()) {
+              servers += server.second->name;
+              servers += " ";
+            }
+            servers += server.second->remote.toStringWithPort();
+          }
+
+          ret << (fmt % name % cache % policy % servers) << endl;
+        }
+        g_outputBuffer=ret.str();
+      }catch(std::exception& e) { g_outputBuffer=e.what(); throw; }
+    });
+
+  g_lua.writeFunction("getPool", [client](const string& poolName) {
+      if (client) {
+        return std::make_shared<ServerPool>();
+      }
+      auto localPools = g_pools.getCopy();
+      std::shared_ptr<ServerPool> pool = createPoolIfNotExists(localPools, poolName);
+      g_pools.setState(localPools);
+      return pool;
+    });
+
+  g_lua.writeFunction("setVerboseHealthChecks", [](bool verbose) { g_verboseHealthChecks=verbose; });
+  g_lua.writeFunction("setStaleCacheEntriesTTL", [](uint32_t ttl) { g_staleCacheEntriesTTL = ttl; });
+
+  g_lua.writeFunction("showBinds", []() {
+      setLuaNoSideEffect();
+      try {
+        ostringstream ret;
+        boost::format fmt("%1$-3d %2$-20.20s %|25t|%3$-8.8s %|35t|%4%" );
+        //             1    2           3            4
+        ret << (fmt % "#" % "Address" % "Protocol" % "Queries" ) << endl;
+
+        size_t counter = 0;
+        for (const auto& front : g_frontends) {
+          ret << (fmt % counter % front->local.toStringWithPort() % (front->udpFD != -1 ? "UDP" : "TCP") % front->queries) << endl;
+          counter++;
+        }
+        g_outputBuffer=ret.str();
+      }catch(std::exception& e) { g_outputBuffer=e.what(); throw; }
+    });
+
+  g_lua.writeFunction("getBind", [](size_t num) {
+      setLuaNoSideEffect();
+      ClientState* ret = nullptr;
+      if(num < g_frontends.size()) {
+        ret=g_frontends[num];
+      }
+      return ret;
+      });
+
+  g_lua.writeFunction("help", [](boost::optional<std::string> command) {
+      setLuaNoSideEffect();
+      g_outputBuffer = "";
+      for (const auto& keyword : g_consoleKeywords) {
+        if (!command) {
+          g_outputBuffer += keyword.toString() + "\n";
+        }
+        else if (keyword.name == command) {
+          g_outputBuffer = keyword.toString() + "\n";
+          return;
+        }
+      }
+      if (command) {
+        g_outputBuffer = "Nothing found for " + *command + "\n";
+      }
+    });
+
+  g_lua.writeFunction("showVersion", []() {
+      setLuaNoSideEffect();
+      g_outputBuffer = "dnsdist " + std::string(VERSION) + "\n";
+    });
+
+#ifdef HAVE_EBPF
+  g_lua.writeFunction("setDefaultBPFFilter", [](std::shared_ptr<BPFFilter> bpf) {
+      if (g_configurationDone) {
+        g_outputBuffer="setDefaultBPFFilter() cannot be used at runtime!\n";
+        return;
+      }
+      g_defaultBPFFilter = bpf;
+    });
+
+  g_lua.writeFunction("registerDynBPFFilter", [](std::shared_ptr<DynBPFFilter> dbpf) {
+      if (dbpf) {
+        g_dynBPFFilters.push_back(dbpf);
+      }
+    });
+
+  g_lua.writeFunction("unregisterDynBPFFilter", [](std::shared_ptr<DynBPFFilter> dbpf) {
+      if (dbpf) {
+        for (auto it = g_dynBPFFilters.begin(); it != g_dynBPFFilters.end(); it++) {
+          if (*it == dbpf) {
+            g_dynBPFFilters.erase(it);
+            break;
+          }
+        }
+      }
+    });
+
+  g_lua.writeFunction("addBPFFilterDynBlocks", [](const map<ComboAddress,int>& m, std::shared_ptr<DynBPFFilter> dynbpf, boost::optional<int> seconds) {
+      setLuaSideEffect();
+      struct timespec until, now;
+      clock_gettime(CLOCK_MONOTONIC, &now);
+      until=now;
+      int actualSeconds = seconds ? *seconds : 10;
+      until.tv_sec += actualSeconds;
+      for(const auto& capair : m) {
+        dynbpf->block(capair.first, until);
+      }
+    });
+
+#endif /* HAVE_EBPF */
+
+  g_lua.writeFunction<std::unordered_map<string,uint64_t>()>("getStatisticsCounters", []() {
+      setLuaNoSideEffect();
+      std::unordered_map<string,uint64_t> res;
+      for(const auto& entry : g_stats.entries) {
+        if(const auto& val = boost::get<DNSDistStats::stat_t*>(&entry.second))
+          res[entry.first] = (*val)->load();
+      }
+      return res;
+    });
+
+  g_lua.writeFunction("includeDirectory", [](const std::string& dirname) {
+      if (g_configurationDone) {
+        errlog("includeDirectory() cannot be used at runtime!");
+        g_outputBuffer="includeDirectory() cannot be used at runtime!\n";
+        return;
+      }
+
+      if (g_included) {
+        errlog("includeDirectory() cannot be used recursively!");
+        g_outputBuffer="includeDirectory() cannot be used recursively!\n";
+        return;
+      }
+
+      g_included = true;
+      struct stat st;
+      if (stat(dirname.c_str(), &st)) {
+        errlog("The included directory %s does not exist!", dirname.c_str());
+        g_outputBuffer="The included directory " + dirname + " does not exist!";
+        return;
+      }
+
+      if (!S_ISDIR(st.st_mode)) {
+        errlog("The included directory %s is not a directory!", dirname.c_str());
+        g_outputBuffer="The included directory " + dirname + " is not a directory!";
+        return;
+      }
+
+      DIR *dirp;
+      struct dirent *ent;
+      std::list<std::string> files;
+      if (!(dirp = opendir(dirname.c_str()))) {
+        errlog("Error opening the included directory %s!", dirname.c_str());
+        g_outputBuffer="Error opening the included directory " + dirname + "!";
+        return;
+      }
+
+      while((ent = readdir(dirp)) != NULL) {
+        if (ent->d_name[0] == '.') {
+          continue;
+        }
+
+        if (boost::ends_with(ent->d_name, ".conf")) {
+          std::ostringstream namebuf;
+          namebuf << dirname.c_str() << "/" << ent->d_name;
+
+          if (stat(namebuf.str().c_str(), &st) || !S_ISREG(st.st_mode)) {
+            continue;
+          }
+
+          files.push_back(namebuf.str());
+        }
+      }
+
+      closedir(dirp);
+      files.sort();
+
+      for (auto file = files.begin(); file != files.end(); ++file) {
+        std::ifstream ifs(*file);
+        if (!ifs) {
+          warnlog("Unable to read configuration from '%s'", *file);
+        } else {
+          vinfolog("Read configuration from '%s'", *file);
+        }
+
+        g_lua.executeCode(ifs);
+      }
+
+      g_included = false;
+    });
+
+  g_lua.writeFunction("setAPIWritable", [](bool writable, boost::optional<std::string> apiConfigDir) {
+      setLuaSideEffect();
+      g_apiReadWrite = writable;
+      if (apiConfigDir) {
+        if (!(*apiConfigDir).empty()) {
+          g_apiConfigDirectory = *apiConfigDir;
+        }
+        else {
+          errlog("The API configuration directory value cannot be empty!");
+          g_outputBuffer="The API configuration directory value cannot be empty!";
+        }
+      }
+    });
+
+  g_lua.writeFunction("setServFailWhenNoServer", [](bool servfail) {
+      setLuaSideEffect();
+      g_servFailOnNoPolicy = servfail;
+    });
+
+  g_lua.writeFunction("setRingBuffersSize", [](size_t capacity) {
+      setLuaSideEffect();
+      if (g_configurationDone) {
+        errlog("setRingBuffersSize() cannot be used at runtime!");
+        g_outputBuffer="setRingBuffersSize() cannot be used at runtime!\n";
+        return;
+      }
+      g_rings.setCapacity(capacity);
+    });
+
+  g_lua.writeFunction("setWHashedPertubation", [](uint32_t pertub) {
+      setLuaSideEffect();
+      g_hashperturb = pertub;
+    });
+
+  g_lua.writeFunction("setTCPUseSinglePipe", [](bool flag) {
+      if (g_configurationDone) {
+        g_outputBuffer="setTCPUseSinglePipe() cannot be used at runtime!\n";
+        return;
+      }
+      setLuaSideEffect();
+      g_useTCPSinglePipe = flag;
+    });
+
+  g_lua.writeFunction("snmpAgent", [](bool enableTraps, boost::optional<std::string> masterSocket) {
+#ifdef HAVE_NET_SNMP
+      if (g_configurationDone) {
+        errlog("snmpAgent() cannot be used at runtime!");
+        g_outputBuffer="snmpAgent() cannot be used at runtime!\n";
+        return;
+      }
+
+      if (g_snmpEnabled) {
+        errlog("snmpAgent() cannot be used twice!");
+        g_outputBuffer="snmpAgent() cannot be used twice!\n";
+        return;
+      }
+
+      g_snmpEnabled = true;
+      g_snmpTrapsEnabled = enableTraps;
+      g_snmpAgent = new DNSDistSNMPAgent("dnsdist", masterSocket ? *masterSocket : std::string());
+#else
+      errlog("NET SNMP support is required to use snmpAgent()");
+      g_outputBuffer="NET SNMP support is required to use snmpAgent()\n";
+#endif /* HAVE_NET_SNMP */
+    });
+
+  g_lua.writeFunction("sendCustomTrap", [](const std::string& str) {
+#ifdef HAVE_NET_SNMP
+      if (g_snmpAgent && g_snmpTrapsEnabled) {
+        g_snmpAgent->sendCustomTrap(str);
+      }
+#endif /* HAVE_NET_SNMP */
+    });
+
+  g_lua.writeFunction("setPoolServerPolicy", [](ServerPolicy policy, string pool) {
+      setLuaSideEffect();
+      auto localPools = g_pools.getCopy();
+      setPoolPolicy(localPools, pool, std::make_shared<ServerPolicy>(policy));
+      g_pools.setState(localPools);
+    });
+
+  g_lua.writeFunction("setPoolServerPolicyLua", [](string name, policyfunc_t policy, string pool) {
+      setLuaSideEffect();
+      auto localPools = g_pools.getCopy();
+      setPoolPolicy(localPools, pool, std::make_shared<ServerPolicy>(ServerPolicy{name, policy}));
+      g_pools.setState(localPools);
+    });
+
+  g_lua.writeFunction("showPoolServerPolicy", [](string pool) {
+      setLuaSideEffect();
+      auto localPools = g_pools.getCopy();
+      auto poolObj = getPool(localPools, pool);
+      if (poolObj->policy == nullptr) {
+        g_outputBuffer=g_policy.getLocal()->name+"\n";
+      } else {
+        g_outputBuffer=poolObj->policy->name+"\n";
+      }
+    });
+
+  g_lua.writeFunction("setTCPDownstreamCleanupInterval", [](uint16_t interval) {
+      setLuaSideEffect();
+      g_downstreamTCPCleanupInterval = interval;
+    });
+
+  g_lua.writeFunction("setConsoleConnectionsLogging", [](bool enabled) {
+      g_logConsoleConnections = enabled;
+    });
+
+  g_lua.writeFunction("setUDPMultipleMessagesVectorSize", [](size_t vSize) {
+      if (g_configurationDone) {
+        errlog("setUDPMultipleMessagesVectorSize() cannot be used at runtime!");
+        g_outputBuffer="setUDPMultipleMessagesVectorSize() cannot be used at runtime!\n";
+        return;
+      }
+#if defined(HAVE_RECVMMSG) && defined(HAVE_SENDMMSG) && defined(MSG_WAITFORONE)
+      setLuaSideEffect();
+      g_udpVectorSize = vSize;
+#else
+      errlog("recvmmsg() support is not available!");
+      g_outputBuffer="recvmmsg support is not available!\n";
+#endif
+    });
+}
+
+vector<std::function<void(void)>> setupLua(bool client, const std::string& config)
+{
+  g_launchWork= new vector<std::function<void(void)>>();
+
+  setupLuaActions();
+  setupLuaConfig(client);
+  setupLuaBindings(client);
+  setupLuaBindingsDNSQuestion();
+  setupLuaInspection();
+  setupLuaRules();
+  setupLuaVars();
+
   std::ifstream ifs(config);
-  if(!ifs) 
+  if(!ifs)
     warnlog("Unable to read configuration from '%s'", config);
   else
     vinfolog("Read configuration from '%s'", config);
 
   g_lua.executeCode(ifs);
 
-  auto ret=*g_launchWork;
+  auto ret = *g_launchWork;
   delete g_launchWork;
-  g_launchWork=0;
+  g_launchWork = nullptr;
   return ret;
 }
