+++ pdns/dnsdistdist/dnsdist-lua-actions.cc
@@ -1,0 +1,1 @@
+../dnsdist-lua-actions.cc
