+++ pdns/dnsdistdist/dnsdist-lua-bindings-dnsquestion.cc
@@ -1,0 +1,1 @@
+../dnsdist-lua-bindings-dnsquestion.cc
