+++ pdns/dnsdistdist/dnsdist-lua-bindings.cc
@@ -1,0 +1,1 @@
+../dnsdist-lua-bindings.cc
