+++ pdns/dnsdistdist/dnsdist-lua-inspection.cc
@@ -1,0 +1,1 @@
+../dnsdist-lua-inspection.cc
