+++ pdns/dnsdistdist/dnsdist-lua-rules.cc
@@ -1,0 +1,1 @@
+../dnsdist-lua-rules.cc
