+++ pdns/dnsdistdist/dnsdist-lua-vars.cc
@@ -1,0 +1,1 @@
+../dnsdist-lua-vars.cc
