--- pdns/dnsdistdist/dnsdist-lua2.cc
@@ -1,1 +1,0 @@
-../dnsdist-lua2.cc
