--- pdns/dnsdistdist/dnsrulactions.hh
@@ -1,1 +1,0 @@
-../dnsrulactions.hh
